Added back Flywheel changes
Due to issues with Github Desktop not performing a 3-way merge correctly
and some merge conflicts, we lost changes to the flywheel.java between
commits 58fdf18 and 829a4ff.
This adds the changes made in commits a68cb9, 32da6e, 9b57c8, 8c710d,
and 51ffc8.
And as it turns out, the rest the changes made in the commits between
58fdf18 and 829a4ff were added in the merge commit (a426345). My guess
it that the flywheel class got overriden with Lucy's version since that
one had a conflict.